+++ LotusRMS.Models/Viewmodels/Purchase/CreatePurchaseDetailVM.cs
@@ -10,10 +10,10 @@ namespace LotusRMS.Models.Viewmodels.Purchase
     {
         public Guid Product_Id { get; set; }
         public string? Product_Name { get; set; }
-        public float Product_Quantity { get; set; }
+        public double Product_Quantity { get; set; }
         public string? Product_Unit { get; set; }
-        public float Product_Rate { get; set; }
-        public float Total {
+        public double Product_Rate { get; set; }
+        public double Total {
             get { return Product_Quantity * Product_Rate; }
             
         }
